+++ examples/settings.js
@@ -55,6 +55,7 @@ let stateIds =
     ];
 export var settings = {
   title: 'React-Dash v0.6.x -- Climate Indices by Year',
+  dashWrapperClass: 'example-dash-wrapper',
   queries: {},
   doFilterRouting: false,
   // start dataResources example