+++ src/components/Autocomplete.js
@@ -3,6 +3,8 @@ import Registry from '../utils/Registry';
 import BaseFilter from './BaseFilter';
 import ReactSelect from './ReactSelect';
 import { makeKey } from '../utils/utils';
+import { omit } from 'lodash';
+import cx from 'classnames/bind';
 
 export default class Autocomplete extends BaseFilter {
   constructor(props) {
@@ -11,15 +13,25 @@ export default class Autocomplete extends BaseFilter {
     if (!this.state.key) {
       this.state.key = makeKey();
     }
-  }  
-  
+  }
+
   render(){
     let val = this.getFilterValue();
-    
-    if (val && !this.props.multi) val = val[0];
-    
+    let props = omit(this.props, 'className');
+    let label = props.label || 'Filter Label';
+    let labelClass = (props.label) ? '' : 'sr-only';
+    let { className } = this.props;
+    let inputProps = {};
+
+    inputProps.id = this.state.key;
+
+    if (val && !props.multi) val = val[0];
+
     return (
-      <ReactSelect.Async value={val} loadOptions={this.loadOptions.bind(this)} disabled={this.isDisabled()} {...this.props} onChange={this.onChange.bind(this)} key={this.state.key} />
+      <div className={cx('autocomplete-filter-container', className)}>
+        <label htmlFor={this.state.key} className={labelClass}>Filter Label</label>
+        <ReactSelect.Async value={val} loadOptions={this.loadOptions.bind(this)} disabled={this.isDisabled()} {...props} onChange={this.onChange.bind(this)} key={this.state.key} inputProps={inputProps} />
+      </div>
     );
   }
 }