+++ src/components/Highlight.js
@@ -14,10 +14,22 @@ export default class Highlight extends BaseComponent {
           { cols.map( (col, i) => {
               return <div className={"highlight-col highlight-cols-" + cols.length} key={i}>
                 { col.rows.map( (row, j) => {
-                  return <div className="highlight-row" key={j}>
-                    <span className="highlight-label">{ row.label }</span>
-                    <span className="highlight-val">{ row.val }</span>
-                  </div>
+                  let labelVal = <span className="highlight-label">{ row.label }</span>;
+                  let rowVal;
+                  
+                  if (row.isLink) {
+                    rowVal = <span className="highlight-val">
+                               <a href={row.val}>{row.val}</a>
+                             </span> 
+                  } else {
+                    rowVal = <span className="highlight-val">{ row.val }</span>  
+                  }
+                  
+                  return (
+                    <div className="highlight-row" key={j}>
+                      {labelVal}{rowVal}
+                    </div>
+                 )
                 }) }
             </div>
           }) }