+++ CHANGELOG.md
@@ -15,6 +15,11 @@
 Card regions can be set as props on the wrapped element, and as a consequence can be set via stateHandlers, allowing us to change the appearance of the card based on component state for wrapped elements.
 - Add stateHandlers
 State handlers fire on component mount and state update. They should be passed an `attr` attribute which tells the stateHandler which component prop to update. If the handler updates a cardVar (card region above or cardClasses narray, etc, the properties will be passed to the wrapping Card element for the wrapped component)
+- Allow accordion regions to open/close with a click.
+- Add accessible headers for card elements.
+- Add accessible filter labels for the autocomplete component.
+- Add support for props.dashWrapperClass in dashboard component
+- Add support for props.isLink in highlight component
 
 0.6.x 2017-3-7-2017 
 -------------------